Workflow and specification divorce
